requirements.txt
yt-dlp
python-telegram-bot>=21.0